+++ mod.ts
@@ -1,4 +1,4 @@
-export * from "./proc.ts";
+export * from "./runner.ts";
 export * from "./runners/proc-group.ts";
 export * from "./runners/handlers/string.ts";
 export * from "./runners/handlers/bytes.ts";