+++ runners/handlers/reader.test.ts
@@ -1,13 +1,13 @@
 import { StringReader } from "../../deps.ts";
 import { assertEquals } from "../../deps-test.ts";
-import { procGroup } from "../proc-group.ts";
+import { group } from "../proc-group.ts";
 import { readerInput } from "./reader.ts";
 import { stringOutput } from "./string.ts";
 
 Deno.test({
   name: "[HAPPY-PATH] I can read stdin from a reader.",
   async fn() {
-    const proc = procGroup();
+    const proc = group();
     try {
       const result = await proc.run(
         readerInput(),